Refactor into components and constants folders && implicit use of index,js as entry
