+++ src/constants/index.js
@@ -1,0 +1,7 @@
+export const DEFAULT_QUERY = 'redux';
+export const DEFAULT_HPP = '100';
+export const PATH_BASE = 'https://hn.algolia.com/api/v1';
+export const PATH_SEARCH = '/search';
+export const PARAM_SEARCH = 'query=';
+export const PARAM_PAGE = 'page=';
+export const PARAM_HPP = 'hitsPerPage=';
